interpretator was connected with interface
