+++ app/src/main/java/com/example/ohana/logic/MathAndLogic.kt
@@ -1,3 +1,5 @@
+package com.example.ohana.logic
+
 val MATH_OPERATORS = arrayOf("+", "-", "*", "/", "//", "%", "**")
 val LOGIC_OPERATORS = arrayOf("<", "<=", ">", ">=", "==", "!=", "&&", "||", "!")
 
@@ -201,117 +203,4 @@ fun solveLogicalExpression(scope: Scope, expression: String): Any {
         }
     }
     return stack.removeLast()
-}
-
-class Scope {
-    val variables: MutableMap<String, Double> = mutableMapOf()
-        get() = field
-    val logs: MutableList<String> = mutableListOf()
-        get() = field
-
-    fun setVariable(name: String, value: Double) {
-        this.variables[name] = value
-    }
-
-    fun setLog(log: String) {
-        logs.add(log)
-    }
-}
-
-open class Block{
-    var id: Int = 0
-    open fun execute() { }
-}
-
-class Body {
-    val _body: MutableList<Block> = mutableListOf(Block())
-        get() = field
-
-    fun appendBlock(block: Block) {
-        _body.add(block)
-    }
-}
-
-class setVariableBlock(val scope: Scope, val name: String, val value: String): Block() {
-    override fun execute() {
-        scope.setVariable(name, solveMathExpression(scope, value))
-    }
-}
-
-class PrintBlock(val scope: Scope, val log: String): Block() {
-    override fun execute() {
-        scope.setLog(solveMathExpression(scope, log).toString())
-    }
-}
-
-class IfBlock(val scope: Scope, val condition: String): Block() {
-    val ifBody = Body()
-
-    override fun execute() {
-        if (solveLogicalExpression(scope, condition) == 1.0) {
-            ifBody._body.forEach {
-                it.execute()
-            }
-        }
-    }
-}
-
-class IfElseBlock(val scope: Scope, val condition: String): Block() {
-    val ifBody = Body()
-    val elseBody = Body()
-
-    override fun execute() {
-        if (solveLogicalExpression(scope, condition) == 1.0) {
-            ifBody._body.forEach {
-                it.execute()
-            }
-        }
-        else {
-            elseBody._body.forEach {
-                it.execute()
-            }
-        }
-    }
-}
-
-class WhileBlock(val scope: Scope, var condition: String): Block() {
-    val body = Body()
-
-    override fun execute() {
-        if (solveLogicalExpression(scope, condition) == 1.0) {
-            body._body.forEach {
-                it.execute()
-            }
-            this.execute()
-        }
-    }
-}
-
-class Interpretator(val scope: Scope) {
-    var program = Body()
-
-    fun launch() {
-        program._body.forEach {
-            it.execute()
-        }
-    }
-}
-
-fun main() {
-    val scope = Scope()
-    val int = Interpretator(scope)
-    val body = Body()
-    val whileBlock = WhileBlock(scope, "a > b")
-
-    whileBlock.body.appendBlock(setVariableBlock(scope, "b", "b + 10"))
-
-    body.appendBlock(setVariableBlock(scope, "a", "81"))
-    body.appendBlock(setVariableBlock(scope, "b", "65"))
-    body.appendBlock(whileBlock)
-
-    int.program = body
-
-    int.launch()
-    print(scope.variables)
-
 }
