--- app/src/main/java/com/example/ohana/sampledata/Parser.kt
@@ -1,4 +1,0 @@
-package com.example.ohana.sampledata
-
-class Parser {
-}
